Conflicts:
	packages/minimongo/minimongo.js
	packages/minimongo/package.js

+++ tools/mongo_runner.js
@@ -243,10 +243,12 @@ exports.launchMongo = function (options) {
       var replSetReady = false;
       var replSetReadyToBeInitiated = false;
       var alreadyInitiatedReplSet = false;
+      var alreadyCalledOnListen = false;
       var maybeCallOnListen = function () {
-        if (listening && replSetReady) {
+        if (listening && replSetReady && !alreadyCalledOnListen) {
           if (createReplSet)
             fs.writeFileSync(portFile, options.port);
+          alreadyCalledOnListen = true;
           onListen();
         }
       };
